refactor:Move SubmitButton and GoBackButton components to atoms folder
